Change references of getline to getline_ncsa and getline_ncsa2
Since NCSA httpd 1.5.2 was released getline was added into libc
and now causes conflicting definition errors during compilation.
getline_ncsa is apparently the same as libc getline, but I still
keep the definitions for historic purposes.

getline_ncsa2 is different (no idea why, originally it was also
called getline), so I changed all definitions of that one to
point to a new name which does not fuck up the compile.